New Functionality and Removed txt file
Some new functionality was added for name recognition and while loops for certain functions. The name recognition is implemented but does not currently function. The txt file was removed since it does not work when the file is within the folder but within my project folder. This file will be reintroduced later.